requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg[binary]
alembic
python-jose
passlib[bcrypt]
bcrypt<4
python-dotenv
pytest
httpx
pytest-cov
